pomelo-llm-server/requirements.txt
transformers
flask
accelerate
bitsandbytes
regex
requests
pyjwt
torch
waitress